Creates new canvas-display package for last of canvas functionality
# Conflicts:
#	bundles/pixi.js/src/index.js
