+++ .travis.yml
@@ -51,7 +51,7 @@ deploy:
     acl: public_read
     region: eu-west-1
     cache_control: "max-age=60"
-    local_dir: bundles/pixi-modern/lib
+    local_dir: bundles/pixi.js-legacy/lib
     upload-dir: "$TRAVIS_BRANCH"
     on:
         all_branches: true
@@ -104,7 +104,7 @@ deploy:
     acl: public_read
     region: eu-west-1
     cache_control: "max-age=2592000"
-    local_dir: bundles/pixi-modern/lib
+    local_dir: bundles/pixi.js-legacy/lib
     upload-dir: "$TRAVIS_BRANCH"
     on:
         all_branches: true
@@ -157,7 +157,7 @@ deploy:
     acl: public_read
     region: eu-west-1
     cache_control: "max-age=1209600"
-    local_dir: bundles/pixi-modern/lib
+    local_dir: bundles/pixi.js-legacy/lib
     upload-dir: "release"
     on:
         all_branches: true
