+++ packages/canvas/canvas-display/README.md
@@ -1,0 +1,13 @@
+# @pixi/canvas-display
+
+## Installation
+
+```bash
+npm install @pixi/canvas-display
+```
+
+## Usage
+
+```js
+import '@pixi/canvas-display';
+```
