+++ packages/canvas/canvas-display/src/index.js
@@ -1,0 +1,2 @@
+import './Container';
+import './DisplayObject';
