+++ packages/canvas/canvas-display/test/.eslintrc.json
@@ -1,0 +1,12 @@
+{
+    "globals": {
+        "expect": false,
+        "assert": false,
+        "sinon": false,
+        "PIXI": false
+    },
+    "rules": {
+        "func-names": 0,
+        "no-unused-expressions": 0
+    }
+}
