+++ packages/interaction/test/InteractionManager.js
@@ -8,6 +8,8 @@ const { InteractionManager } = require('../');
 const { CanvasGraphicsRenderer } = require('@pixi/canvas-graphics');
 const { CanvasSpriteRenderer } = require('@pixi/canvas-sprite');
 
+require('@pixi/canvas-display');
+
 CanvasRenderer.registerPlugin('interaction', InteractionManager);
 CanvasRenderer.registerPlugin('graphics', CanvasGraphicsRenderer);
 CanvasRenderer.registerPlugin('sprite', CanvasSpriteRenderer);
